skills/master-orchestrator/requirements.txt
# MasterOrchestrator Dependencies
# Python 3.8+ required

# Core dependencies
pyyaml>=6.0.1          # YAML configuration file parsing
chardet>=5.2.0         # Character encoding detection for event parsing

# Optional dependencies for extended features
aiohttp>=3.9.0         # Async HTTP client for AduibClient (optional)

# Development dependencies (optional)
# pytest>=7.4.0        # Testing framework
# pytest-asyncio>=0.21.0  # Async testing support
# black>=23.0.0        # Code formatting
# mypy>=1.7.0          # Type checking

# External tools (install separately)
# npm install -g memex-cli  # Required for backend orchestration
